refactor: update component imports and detail screen sections
- Move DetailExpandableText to presentation.components.expandable_text
- Switch DroppedItemsSection to accept icon lambda
- Replace HorizontalPagerWithHeaderData with SectionHeaderData
- Remove beyondViewportPageCount from HorizontalPager
- Update PassiveCreatureDetailScreen icon to Lucide.Trophy
