+++ app/src/main/java/com/rabbitv/valheimviki/presentation/detail/tool/ToolDetailScreen.kt
@@ -44,7 +44,7 @@ import com.rabbitv.valheimviki.domain.model.upgrader.MaterialUpgrade
 import com.rabbitv.valheimviki.navigation.BuildingDetailDestination
 import com.rabbitv.valheimviki.navigation.DetailDestination
 import com.rabbitv.valheimviki.navigation.NavigationHelper
-import com.rabbitv.valheimviki.presentation.components.DetailExpandableText
+import com.rabbitv.valheimviki.presentation.components.expandable_text.DetailExpandableText
 import com.rabbitv.valheimviki.presentation.components.bg_image.BgImage
 import com.rabbitv.valheimviki.presentation.components.button.AnimatedBackButton
 import com.rabbitv.valheimviki.presentation.components.button.FavoriteButton
@@ -55,9 +55,9 @@ import com.rabbitv.valheimviki.presentation.components.flow_row.flow_as_grid.Two
 import com.rabbitv.valheimviki.presentation.components.grid.grid_item.CustomItemCard
 import com.rabbitv.valheimviki.presentation.components.horizontal_pager.HorizontalPagerData
 import com.rabbitv.valheimviki.presentation.components.horizontal_pager.HorizontalPagerSection
-import com.rabbitv.valheimviki.presentation.components.horizontal_pager.HorizontalPagerWithHeaderData
 import com.rabbitv.valheimviki.presentation.components.images.FramedImage
 import com.rabbitv.valheimviki.presentation.components.section_header.SectionHeader
+import com.rabbitv.valheimviki.presentation.components.section_header.SectionHeaderData
 import com.rabbitv.valheimviki.presentation.components.trident_divider.TridentsDividedRow
 import com.rabbitv.valheimviki.presentation.detail.tool.model.ToolDetailUiState
 import com.rabbitv.valheimviki.presentation.detail.tool.viewmodel.ToolDetailViewModel
@@ -161,15 +161,11 @@ fun ToolDetailContent(
 								.padding(horizontal = BODY_CONTENT_PADDING.dp)
 						) {
 							SectionHeader(
-								data = HorizontalPagerWithHeaderData(
+								data = SectionHeaderData(
 									title = "Crafting Ingredients",
 									subTitle = "Items required to craft this item",
-									icon = Lucide.Hammer,
-									iconRotationDegrees = 0f,
-									contentScale = ContentScale.Crop,
-									starLevelIndex = 0,
-								),
-								modifier = Modifier,
+									icon = Lucide.Hammer
+								)
 							)
 						}
